+++ common/model/src/main/java/is/codion/common/model/condition/DefaultTableConditionModel.java
@@ -28,7 +28,6 @@ import is.codion.common.reactive.value.ValueSet;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
-import java.util.Optional;
 import java.util.Set;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
@@ -81,11 +80,6 @@ final class DefaultTableConditionModel<C> implements TableConditionModel<C> {
 		}
 
 		return condition;
-	}
-
-	@Override
-	public <T> Optional<ConditionModel<T>> optional(C identifier) {
-		return Optional.ofNullable((ConditionModel<T>) conditions.get(requireNonNull(identifier)));
 	}
 
 	@Override